Move PostManger and PostQuerySet to board/models.py, Add search flag option
